Update Challenge 2  Conversion de la température.c

+++ Day01/Exercices sur les Variables/Challenge 2  Conversion de la température.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int T ;
+    float T ;
     float K;
     printf("Vuillez Entrez la temperature qui vou voullez changer :");
     scanf("%d",&T);
